Reorganize test
